fix: remove useless meid
