+++ x/megroup/types/expected_keepers.go
@@ -11,17 +11,6 @@ import (
 )
 
 type StakingKeeper interface {
-	// Methods imported from staking should be defined here
-
-	// FIXME: wait wstaking keep and types.Meid ;Temporarily use MockMeid instead of MEID and MockRegion instead types.Region
-	// GetMeid(ctx sdk.Context, account string) (val types.Meid, found bool)
-	GetMeid(ctx sdk.Context, account string) (val wstakingtypes.Meid, found bool)
-	// SetMeid(ctx sdk.Context, meid types.Meid)
-	SetMeid(ctx sdk.Context, meid wstakingtypes.Meid)
-
-	// GetRegion(ctx sdk.Context, regionId string) (val types.Region, found bool)
-	//	GetRegion(ctx sdk.Context, regionId string) (val mock.MockRegion, found bool)
-
 	GetRegion(ctx sdk.Context, regionId string) (region wstakingtypes.Region, found bool)
 }
